% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMatrix.R
\name{gMatrix}
\alias{gMatrix}
\title{Calculate the \eqn{g_{ij}} values.}
\usage{
gMatrix(Y)
}
\arguments{
\item{Y}{\eqn{N-by-q} matrix. The sample dataset of Y. Each row represents one observation.}
}
\value{
A list containing the following components will be returned:
\describe{
\item{gM}{\eqn{N(N-1)/2-by-1} matrix. The g-Matrix of Y.}
\item{sum4}{The sum of distance matrix of Y.}
}
}
\description{
\code{gMatrix} is used to calulate all the \eqn{g_{ij}} values
for \eqn{1\le i<j\le N}, where \eqn{N} is the sample size of the data.
}
\examples{
library(MASS)
Num = 40
p = 3
q = 2
mu1 = rep(0,p)
sigma1 = diag(p)
mu2 = rep(0,q)
sigma2 = diag(q)
X = mvrnorm(Num,mu1,sigma1)
Y = mvrnorm(Num,mu2,sigma2)
SUM_X = rowSums(X)
Y[,1] = SUM_X+0.01*Y[,1]
Y[,2] = (SUM_X)^2+0.01*Y[,2]
gMatrix_Y <- gMatrix(Y)
gMatrix_X <- gMatrix(X)
}
