% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxDCOV.R
\name{maxDCOV}
\alias{maxDCOV}
\title{Maximize the distance covariance of \eqn{Xu} and \eqn{Y} under certain constraint.}
\usage{
maxDCOV(
  Xd,
  gM,
  u0 = c(1, rep(0, ncol(Xd) - 1)),
  e = 10^-3,
  xi = 200,
  psi = 10,
  rho = 1000,
  omega = 10,
  abs.epsilon = 10^-3,
  rel.epsilon = 10^-3
)
}
\arguments{
\item{Xd}{\eqn{N(N-1)/2-by-p} matrix. The matrix of the pairwise difference of X.
Each row represents one observation. It can be obtained by function \code{Xdiff}.}

\item{gM}{\eqn{N(N-1)/2-by-q} matrix. The g-Matrix of Y.
It can be obtained by function \code{gMatrix}.}

\item{u0}{The initial guess of the solution. The default value is (1,0,...,0).}

\item{e}{The tolerance of two values calculated by adjacent two steps by Difference of Convex Algorithm (DCA).
If the two values is smaller than e, we terminate the loop of DCA.
The default value is 10^-3.}

\item{xi}{One of the augmented Lagrangian parameters.
It should be nonnegative as well as bigger than psi.
The default value is 200.}

\item{psi}{The other augmented Lagrangian parameters.
The default value is 10.}

\item{rho}{The augmented Lagrangian parameter in subproblem.
The default value is 1000.}

\item{omega}{The increment of xi. It should be bigger than 1.
The default value is 10.}

\item{abs.epsilon}{The absolute tolerance in ADMM.
The default value is 10^-3.}

\item{rel.epsilon}{The relative tolerance in ADMM.
The default value is 10^-3.}
}
\value{
The solution \eqn{u}.
}
\description{
\code{maxDCOV} is used to solve the non-convex optimization problem of
maximizing the distance covariance of \eqn{Xu} and \eqn{Y} under the constraint that
the Euclidean norm of \eqn{u} is 1.
}
\examples{
library(MASS)
Num = 40
p = 3
q = 2
mu1 = rep(0,p)
sigma1 = diag(p)
mu2 = rep(0,q)
sigma2 = diag(q)
X = mvrnorm(Num,mu1,sigma1)
Y = mvrnorm(Num,mu2,sigma2)
SUM_X = rowSums(X)
Y[,1] = SUM_X+0.01*Y[,1]
Y[,2] = (SUM_X)^2+0.01*Y[,2]
X_diff <- Xdiff(X)
gM_Y <- gMatrix(Y)
g_Y <- gM_Y$gM
\donttest{
maxDCOV(X_diff,g_Y)
}
}
